// backend/src/models/reports/interfaces/report.interface.ts
export interface IReport {
  email: null | string;
  firstName: null | string;
  lastName: null | string;
  preferredName: null | string;
  photoUrl: null | string;
  gender: null | string;
  country: null | string;
  religion: null | string;
  mobile: null | string;
  status: boolean;
  situation: string;
}